Refactors JavaScript to be easier to follow and harder to screw up.
Follows the model in Elm where there is a state mutation function holds business logic, and then you have a render function that knows how to render any given state, but the two are kept separate.  This also makes testing notably easier as we can just test the state transition function without needing a browser/DOM available.

Adds some type definitions.

Hooks up to embedded web3.  This means the app no longer works without MetaMask or similar running.  No need to be logged in, but `web3.currentProvider` needs to be on the page.