Iterating pretty hard on the backing JavaScript.
As the JS continues to grow, I continue to iterate on it to keep it manageable and auditable.  To this end, I have enabled type checking on the JS files via TypeScript's JS type checking feature.  This required fixing up a lot of JSDoc comments and adding a type definition file but it caught a number of type errors (all since fixed).  I have also made almost all objects immutable, and isolated the mutable state into a few controlled locations.

Unrelated to technical debt reduction, I added some scaffolding for the CDP closure UI.  I have not yet applied the ideas @epheph and I discussed around presenting closure as a stepped slider.

Finally, I added a tiny bit of CSS to make things not _quite_ as ugly as they were.