Further iteration on the JavaScript.
Splits out the periodic updates from the blockchain into a poller class whose only job is to regularly poll for updates.

Groups the various Ethereum Client functions and hides some of them so future developers don't try to use them directly and instead create proper accessor functions.

Moves around some validation so it occurs closer to the wire.

Creates a ContractAddresses class that deals with the address for each of the various contracts we work with.

State manager can now be subscribed to so other classes can do any logic they need when state changes.  This is most useful for the presentor, but CDP opener also benefits from this for deciding if it needs to cascade an update to the DAI sale proceeds.