Iterating pretty hard on the backing JavaScript. (#22)
* Iterating pretty hard on the backing JavaScript.

As the JS continues to grow, I continue to iterate on it to keep it manageable and auditable.  To this end, I have enabled type checking on the JS files via TypeScript's JS type checking feature.  This required fixing up a lot of JSDoc comments and adding a type definition file but it caught a number of type errors (all since fixed).  I have also made almost all objects immutable, and isolated the mutable state into a few controlled locations.

Unrelated to technical debt reduction, I added some scaffolding for the CDP closure UI.  I have not yet applied the ideas @epheph and I discussed around presenting closure as a stepped slider.

Finally, I added a tiny bit of CSS to make things not _quite_ as ugly as they were.

* Further iteration on the JavaScript.

Splits out the periodic updates from the blockchain into a poller class whose only job is to regularly poll for updates.

Groups the various Ethereum Client functions and hides some of them so future developers don't try to use them directly and instead create proper accessor functions.

Moves around some validation so it occurs closer to the wire.

Creates a ContractAddresses class that deals with the address for each of the various contracts we work with.

State manager can now be subscribed to so other classes can do any logic they need when state changes.  This is most useful for the presentor, but CDP opener also benefits from this for deciding if it needs to cascade an update to the DAI sale proceeds.

* Rearranges some classes and makes ContractAddresses depend on state.

By depending on state, we can make the calls sync, and not hammer the node unnecessarily.