Turns on TypeScript strict mode and handles insufficient order book depth.
Enabling TS strict mode is uninteresting, and most of the changes in this PR are just addressing the issues that came up or fixing bugs found by strict mode (there were a couple minor ones).

If the Maker function fails, we interpret this to mean there isn't enough order book depth.  This isn't exactly true, and one day we should find a better solution but for now it is a "good enough" correlation.  When that happens, we assign `'insufficient depth'`to the `estimatedPriceOfEthInDai` variable and use that to determine whether or not we should display a warning sign and tooltip to the user.  This solution works out, but it does require handling that string everywhere `estimatedPriceOfEthInDai` is referenced.