Creates initial contract and sets up deployment scripts.
Compilation and deployment are intentionally kept simple since we only have one contract to deploy.

The initial contract is just to make sure compilation and deployment worked.

Next step is to turn the contract into a mock contract so we can test the UI against a local instantseal node with faux data.