Splits client and contracts into separate project subfolders.
This change is mostly just file moving and the associated changes that fell out from that.

The only other change of significance is the introduction of network 17, which is the InstantSeal network used in the docker-compose stuff, and the addition of LiquidLong contracts to the client (unused so far).

Also updated the Readme.