Hooks up new functions to UI.
Added a new method to the contracts, mostly just a copy of the previous one.

Renamed some variables and methods in the contracts to try to be more clear.  The most noticeable one is in line with the comment that all `pay` and `buy` are from the perspective of the caller/taker.  This required swapping one of the variables which was actually from the perspective of the maker (thus causing me great confusion).

Fixes a bug in the contracts where we were multiplying instead of dividing when converting at a particular exchange rate.