Creates a new client-library.
This client-library will be used by the UI and will handle all of the complexity around integrating with Ethereum.

This is mostly scaffolding to get everything hooked up and a single test to make sure it works (it does!)

Also fixes a bug from previous PR that was introduced when applying feedback.