Reorganizes project.
Pulls tests out into a separate package so that the tests can be sure they are testing the whole NPM package, rather than the just the code.  This reduces the likelihood of publishing a package that isn't setup correctly.  It also makes it so the library more correctly presents its dependencies, separate from the tests.