Adds sideEffects bool to package.json for WebPack.
This tells WebPack that it can take certain optimizations with the knowledge that the act of importing this module will not have side effects, thus can be done once, or multiple times without problems and without regard to current context.

Also added explicit `types`, even though the current value is the default, for clarity.