Provides transaction with gas limit of our choosing.
Note that we'll follow this set of rules for _all_ gas estimations, which means everything on-chain costs between 250k and 5M and is 1.3x the node's estimated price.  In production, I believe the only transactions that happen on chain are openCdp tranasctions, so this is fine for now, but we may want to revisit this in the future if we start doing more on-chain stuff.