Changes the way Ethers importing is done to make library work in browser.
The Ethers library has two sets of outputs, one that lives in the root of the project and one that lives in `dist` subdirectory.  The `dist` subdirectory is compiled targeting browser, while the rest is compiled targeting NodeJS.  When you do an import of a library using `import { ... } from 'library/path'` it ignores the `package.json` paths and accesses the library directly at that path.  This means that if you do `import { ... } from 'ethers/utils'` it will always be pulling in the NodeJS version, even when you are targeting the browser.

To complicate matters, TSC seemse to be unhappy with the way `ethers` exports are setup and it complains if a dependency of this library tries to do `import { utils } from 'ethers'` and fulfill a type requirement in the library where it also has a similar import.  To resolve this, the library now re-exports the Ethers pieces that are necessary to create a custom Provider.  This should only be necessary for testing/mocking things, so end-userse shouldn't ever see or notice this change.

The important high-level takeaway from this change is that the library is now friendly to being used in the browser.  It has been tested against a previously failing Angular app (which bundles with Webpack under the hood) and everything is now working.