Library now outputs both commonjs and es2015 modules.
Major version bump since we are moving where the filese are and some workflows may break as a result.

The `.js` suffix in the TS files is an unfortunate consequence of ES modules requiring full filenames and TSC not having an option to add the extension as part of the compile process.  It works and is supported, even though it is dumb.

The changes to the TimeoutScheduler are because previously we were depending on NodeJS specific types and we got lucky that it happened to work in the browser as well.  While the code didn't change, the type checker now is more correctly checking the type.  At runtime the task ID will be either always `number` or always `NodeJS.Timer` depending on environment.

Also updated local environment and Dockerfile to node LTS (v10).