Removes default values from polled values.
These defaults made coding easier because we had a `number` instead of `number | null` but they prevented us from differentiating between "loading" and "loaded".  This change is breaking (hence the major version bump) but I believe it is the right move, and I'm not sure why I decided to not do this in the first place.

The recursive-fs update just makes it so the CLI doesn't print an empty line on success or an error on delete when the directory doesn't exist (QoL improvements).

The new tsconfig.json is just so VS Code is capable of processing the files, since it will _only_ use a file named `tsconfig.json`.